[feat] : JPA 매핑 에러

+++ backend/src/main/java/com/ssafy/fullerting/badge/model/entity/Badge.java
@@ -8,7 +8,7 @@ public class Badge {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int badgeId;
+    private Long badgeId;
 
     @Column(name = "badge_name", nullable = false, length = 20)
     private String name;
